+++ backend/app/oauth.py
@@ -6,8 +6,8 @@ from fastapi.responses import RedirectResponse
 import datetime
 
 sys.path.append(os.environ['OPENETL_HOME'])
-from utils.main_api_class import OAuth2Client
-from utils.database_utils import DatabaseUtils
+from openetl_utils.main_api_class import OAuth2Client
+from openetl_utils.database_utils import DatabaseUtils
 
 
 router = APIRouter(prefix="/auth", tags=["auth"])
