+++ connectors/api/gsheet.py
@@ -2,11 +2,11 @@ import sys
 import os
 import pandas as pd
 from sqlalchemy import create_engine
-from utils import AuthType
+from openetl_utils import AuthType
 
 sys.path.append(os.environ['OPENETL_HOME'])
 
-from utils.main_db_class import DB
+from openetl_utils.main_db_class import DB
 
 
 class Connector(DB):
