+++ openetl_utils/enums.py
